Fix several warnings

+++ components/renderer/base_renderer.hpp
@@ -22,9 +22,13 @@ public:
     virtual bool exportScene(std::string fileName)=0;
     virtual int getSelectedEntityId()=0;
 
-    MFMath::Vec3 getCameraPosition() { double x,y,z,ya,pi,ro; getCameraPositionRotation(x,y,z,ya,pi,ro); return MFMath::Vec3(x,y,z); };
+    MFMath::Vec3 getCameraPosition() 
+    { 
+        double x,y,z,ya,pi,ro; getCameraPositionRotation(x,y,z,ya,pi,ro);
+        return MFMath::Vec3((float)x, (float)y, (float)z);
+    };
 
-protected:
+  protected:
     bool mDone;
 };
